bfloat

+++ model_predictions_generation.py
@@ -40,7 +40,7 @@ torch.cuda.empty_cache()
 print('PREPROCESSING DATA...')
 DATASET_CHEKPOINT="ferrazzipietro/e3c-sentences" 
 TRAIN_LAYER="en.layer1"
-adapters_list = generate_adapters_list(log_name_training, appendix=appendix, training_type=training_type)
+adapters_list = generate_adapters_list(log_name_training, appendix=appendix, training_type=training_type, dtype=dtype)
 if training_type != 'NoLora':
     peft_config = PeftConfig.from_pretrained(adapters_list[0], token = HF_TOKEN_WRITE)
     BASE_MODEL_CHECKPOINT = peft_config.base_model_name_or_path
